pyproject.toml
[project]
name = "style-experiment"
version = "0.1.0"
description = "A comparison of different strategies for affecting the style of an LLM's output."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "datasets>=3.5.0",
    "ipykernel>=6.29.5",
    "loam-iiif>=0.1.3",
    "nltk>=3.9.1",
    "peft>=0.15.2",
    "pillow>=11.2.1",
    "scikit-learn>=1.6.1",
    "tensorboardx>=2.6.2.2",
    "transformers>=4.51.3",
    "trl>=0.16.1",
    "yaspin>=3.1.0",
]
